1.removed access key 2. moved services and routes to separate folders
