+++ src/app/app.module.ts
@@ -3,11 +3,11 @@ import { NgModule } from '@angular/core';
 
 import { AppComponent } from './app.component';
 import { WeatherComponent } from './weather/weather.component';
-import { allAppRoutes } from './routes';
+import { allAppRoutes } from './routes/routes';
 import { RouterModule } from '@angular/router';
 import { ReactiveFormsModule } from '@angular/forms';
 import { HttpClientModule } from '@angular/common/http';
-import { ApixuService } from './apixu.service';
+import { ApixuService } from './services/APIXU/apixu.service';
 import { WeatherDetailsComponent } from './weather-details/weather-details.component';
 
 @NgModule({
