+++ src/app/weather-details/weather-details.component.ts
@@ -1,5 +1,4 @@
 import { Component, OnInit, Input } from '@angular/core';
-import { ApixuService } from '../apixu.service';
 
 
 @Component({
