+++ src/app/weather/weather.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { FormBuilder, FormGroup } from '@angular/forms';
-import { ApixuService } from '../apixu.service';
+import { ApixuService } from '../services/APIXU/apixu.service';
 
 @Component({
   selector: 'app-weather',
